requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
ConfigArgParse==1.7.1
Deprecated==1.3.1
fastapi==0.129.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.3.0
limits==5.8.0
locust==2.43.3
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.0
pluggy==1.6.0
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
pywin32==311
pyzmq==27.1.0
requests==2.32.5
simple-websocket==1.1.0
slowapi==0.1.9
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websocket-client==1.9.0
Werkzeug==3.1.5
wrapt==2.1.1
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
